requirements.txt
fastapi==0.112.1
sqlmodel==0.0.21
uvicorn==0.30.6
strawberry-graphql==0.237.3